files/requirements.txt
# MacroSignalTool — Python dependencies
# pip install -r requirements.txt --break-system-packages

# ── Core AI ───────────────────────────────────────────────────────────────────
anthropic>=0.40.0

# ── News ingestion (Phase 2) ──────────────────────────────────────────────────
feedparser>=6.0.10
requests>=2.31.0

# ── Market data (Phase 2, 7) ──────────────────────────────────────────────────
yfinance>=0.2.40

# ── Web framework (Phase 4) ───────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0

# ── Scheduler (Phase 4, 6) ────────────────────────────────────────────────────
apscheduler>=3.10.4

# ── Data processing (Phase 2, 7) ─────────────────────────────────────────────
pandas>=2.2.0
numpy>=1.26.0

# ── Telegram bot (Phase 6.1) ──────────────────────────────────────────────────
# Per TelegramNotifier (send-only, no polling): solo aiohttp
aiohttp>=3.9.0
# Per MacroSignalBot (comandi interattivi, opzionale):
python-telegram-bot>=21.0

# ── Instagram / Phase 8 ───────────────────────────────────────────────────────
# Playwright per rendering HTML → PNG (slide carosello)
# Dopo pip install, eseguire: playwright install chromium
playwright>=1.44.0

# Timezone support per APScheduler (carosello alle 09:00 CET)
pytz>=2024.1

# ── Utilities ─────────────────────────────────────────────────────────────────
python-dotenv>=1.0.0
python-dateutil>=2.9.0
